fix: move isDev to a independent file
