+++ src/layouts/Header.tsx
@@ -8,9 +8,10 @@ import { Link } from 'react-router-dom'
 import CopyTooltip from '~/components/CopyTooltip'
 import Account from '~/layouts/Account'
 import rootStore from '~/store'
-import { isDev, shortenHash } from '~/utils'
+import { shortenHash } from '~/utils'
 import { mojoToXch } from '~/utils/CoinConverter'
 import { chains } from '~/utils/constants'
+import { isDev } from '~/utils/env'
 import CopyIcon from '~icons/hoogii/copy.jsx'
 import EditIcon from '~icons/hoogii/edit.jsx'
 import InfoIcon from '~icons/hoogii/info.jsx'
