+++ src/utils/index.ts
@@ -2,11 +2,7 @@ import { hash256 } from '@rigidity/bls-signatures'
 import { Program } from '@rigidity/clvm'
 import bcrypt from 'bcryptjs'
 
-import { ModeEnum } from '~/types'
 
-// export const isDev = import.meta.env.DEV
-
-// export const isExtension = import.meta.env.MODE !== ModeEnum.development
 
 export const shortenHash = (
     hash: string,
